@@ -2,21 +2,17 @@ package com.example.my_book_shop_app.struct.author;
 
 import com.example.my_book_shop_app.struct.book.Book;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Table(name = "authors")
-@ApiModel(description = "data model of author entity")
 public class Author {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @ApiModelProperty("author id generated by db")
     private int id;
 
     @Column(columnDefinition = "VARCHAR(255)")
@@ -26,7 +22,6 @@ public class Author {
     private String slug;
 
     @Column(columnDefinition = "VARCHAR(255) NOT NULL")
-    @ApiModelProperty("The name of author")
     private String name;
 
     @Column(columnDefinition = "TEXT")
