@@ -10,54 +10,42 @@ import com.example.my_book_shop_app.struct.tags.TagsEntity;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
 @Entity
 @Table(name = "books")
-@ApiModel(description = "entity representation a book")
 public class Book {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @ApiModelProperty("id generated by db automatically")
     private int id;
 
     @Column(columnDefinition = "DATE NOT NULL")
-    @ApiModelProperty("date of book publication")
     private Date pubDate;
 
     @Column(columnDefinition = "SMALLINT NOT NULL")
-    @ApiModelProperty("if isBestseller = 1 so the book is considered to be bestseller and if 0 the book is not a bestseller")
     private byte isBestseller;
 
     @Column(columnDefinition = "VARCHAR(255) NOT NULL")
-    @ApiModelProperty("mnemonical identity sequence of characters")
     private String slug;
 
     @Column(columnDefinition = "VARCHAR(255) NOT NULL")
-    @ApiModelProperty("book title")
     private String title;
 
     @Column(columnDefinition = "VARCHAR(255)")
-    @ApiModelProperty("book url")
     private String image;
 
     @Column(columnDefinition = "TEXT")
-    @ApiModelProperty("book description text")
     private String description;
 
     @Column(columnDefinition = "INT NOT NULL")
-    @ApiModelProperty("old book price")
     private int price;
 
     @Column(columnDefinition = "SMALLINT NOT NULL DEFAULT 0")
-    @ApiModelProperty("discount value for book")
     private byte discount;
 
     @ManyToOne()
