@@ -3,7 +3,7 @@ package com.example.my_book_shop_app.struct.external_api;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Data
