@@ -2,8 +2,14 @@ package com.example.my_book_shop_app.struct.tags;
 
 import com.example.my_book_shop_app.struct.book.Book;
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.Table;
 
-import javax.persistence.*;
 import java.util.List;
 
 @Entity
